+++ package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "packageManager": "pnpm@10.28.1",
   "scripts": {
-    "dev": "vite",
-    "start": "vite",
+    "dev": "vite --port 3000",
+    "start": "vite --port 3000",
     "build": "vite build",
     "preview": "vite preview",
     "typecheck": "tsc --noEmit",
