--- src/App.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-import { Reheboam } from "./components/mainsection/reheboam";
-
-import "./App.css";
-
-export const App = (): React.JSX.Element => {
-  return <Reheboam />;
-};
