+++ src/app.tsx
@@ -1,0 +1,7 @@
+import { Reheboam } from "./components/mainsection/reheboam";
+
+import "./app.css";
+
+export const App = () => {
+  return <Reheboam />;
+};
