+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 
 import "./index.css";
 
-import { App } from "./App";
+import { App } from "./app";
 
 const rootElement = document.getElementById("root");
